fix some validate when check

+++ controller/food.controller.js
@@ -11,13 +11,17 @@ var salt = bcrypt.genSaltSync(10);
 
 module.exports.create = async function (req, res, next) {
     try {
+        if (req.role == "housewife") {
+            const result = await cloudinary.v2.uploader.upload(req.file.path)
+            req.body.image = result.url;
+            req.body.image_id = result.public_id;
+            //create food
+            var food = await Food.create(req.body);
+            res.status(200).send(food);
+        } else {
+            res.status(401).send({ message: "Không có quyền thêm sản phẩm" });
+        }
 
-        const result = await cloudinary.v2.uploader.upload(req.file.path)
-        req.body.image = result.url;
-        req.body.image_id = result.public_id;
-        //create food
-        var food = await Food.create(req.body);
-        res.status(200).send(food);
     } catch (err) {
         console.log(err.message);
         next(err.message)
@@ -37,10 +41,15 @@ module.exports.getFoods = async function (req, res, next) {
 
 module.exports.delete = async function (req, res, next) {
     try {
-        console.log(req.body);
-        await cloudinary.v2.uploader.destroy(req.params.image_id);
-        var food = await Food.deleteOne({ '_id': req.params.id });
-        res.send({ "message": "Xóa thành công" });
+        if (req.role == "admin") {
+            await cloudinary.v2.uploader.destroy(req.params.image_id);
+            var food = await Food.deleteOne({ '_id': req.params.id });
+            res.send({ "message": "Xóa thành công" });
+        } else {
+            res.status(401).send({ message: "Không có quyền xóa sản phẩm" });
+
+        }
+
     } catch (err) {
         next(err.message)
     }
@@ -48,13 +57,18 @@ module.exports.delete = async function (req, res, next) {
 
 module.exports.checkOk = async function (req, res, next) {
     try {
-        var food = await Food.updateOne({ _id: req.params.id }, {
-            $set: {
-                isCkeck: true
-            }
-        });
-        food.message = "Update thành công"
-        res.json(food);
+        if (req.role == "admin") {
+            var food = await Food.updateOne({ _id: req.params.id }, {
+                $set: {
+                    isCkeck: true
+                }
+            });
+            food.message = "Update thành công"
+            res.json(food);
+        } else {
+            res.status(401).send({ message: "Không có quyền update sản phẩm" });
+        }
+
 
     } catch (err) {
         next(err.message)
